requirements.txt
blessed==1.18.0
botocore==1.20.72
Django==3.2.2
djangorestframework==3.12.4
notion-client
numpy
psycopg2>=2.7.7
python-jose>=3.2.0
python-dotenv>=0.17.1
pytest-django
pytest-xdist
PyYAML
requests>=2.25.1
slack_sdk
social-auth-app-django>=4.0.0
whitenoise>=5.2.0
